+++ docsrc/source/changelog.rst
@@ -17,6 +17,10 @@ Features/Updates:
 - Add smaller dataset export
 - Fix unique cluster IDs
 - Add feature for workflows
+- Add `operate` function to run on each cluster
+- Add a way to create centroids if they do not exist using the `create_centroids` function
+- Fix metadata insertion eperience
+- Fix community detection to return clusterops object
 
 Backend:
 - Major folder refactor -> official renaming of ops to workflows in certain areas